Remove newNavigation feature toggle + old code
