+++ conf/defaults.ini
@@ -1184,9 +1184,6 @@ commandPalette = true
 # Use dynamic labels in CloudWatch datasource
 cloudWatchDynamicLabels = true
 
-# New expandable navigation
-newNavigation = true
-
 # feature1 = true
 # feature2 = false
