+++ packages/grafana-data/src/types/featureToggles.gen.ts
@@ -32,7 +32,6 @@ export interface FeatureToggles {
   prometheus_azure_auth?: boolean;
   prometheusAzureOverrideAudience?: boolean;
   influxdbBackendMigration?: boolean;
-  newNavigation?: boolean;
   showFeatureFlagsInUI?: boolean;
   publicDashboards?: boolean;
   lokiLive?: boolean;
