+++ pkg/api/navlinks/navlinks.go
@@ -2,7 +2,7 @@ package navlinks
 
 import "github.com/grafana/grafana/pkg/api/dtos"
 
-func GetServerAdminNode(children []*dtos.NavLink, navSection string) *dtos.NavLink {
+func GetServerAdminNode(children []*dtos.NavLink) *dtos.NavLink {
 	url := ""
 	if len(children) > 0 {
 		url = children[0].Url
@@ -15,7 +15,7 @@ func GetServerAdminNode(children []*dtos.NavLink, navSection string) *dtos.NavLi
 		Icon:         "shield",
 		Url:          url,
 		SortWeight:   dtos.WeightAdmin,
-		Section:      navSection,
+		Section:      dtos.NavSectionConfig,
 		Children:     children,
 	}
 }
