+++ public/app/core/components/NavBar/NavBar.test.tsx
@@ -36,6 +36,17 @@ const setup = () => {
 };
 
 describe('Render', () => {
+  beforeEach(() => {
+    // IntersectionObserver isn't available in test environment
+    const mockIntersectionObserver = jest.fn();
+    mockIntersectionObserver.mockReturnValue({
+      observe: () => null,
+      unobserve: () => null,
+      disconnect: () => null,
+    });
+    window.IntersectionObserver = mockIntersectionObserver;
+  });
+
   it('should render component', async () => {
     setup();
     const sidemenu = await screen.findByTestId('sidemenu');
