+++ public/app/core/components/NavBar/NavBar.tsx
@@ -1,14 +1,14 @@
 import { css, cx } from '@emotion/css';
+import { FocusScope } from '@react-aria/focus';
 import { cloneDeep } from 'lodash';
 import React, { useState } from 'react';
-import { connect, ConnectedProps } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { useLocation } from 'react-router-dom';
 
 import { GrafanaTheme2, NavModelItem, NavSection } from '@grafana/data';
-import { locationService } from '@grafana/runtime';
-import { Icon, IconName, useTheme2 } from '@grafana/ui';
+import { config, locationService, reportInteraction } from '@grafana/runtime';
+import { Icon, useTheme2 } from '@grafana/ui';
 import { Branding } from 'app/core/components/Branding/Branding';
-import config from 'app/core/config';
 import { getKioskMode } from 'app/core/navigation/kiosk';
 import { KioskMode, StoreState } from 'app/types';
 
@@ -17,169 +17,269 @@ import { OrgSwitcher } from '../OrgSwitcher';
 import NavBarItem from './NavBarItem';
 import { NavBarItemWithoutMenu } from './NavBarItemWithoutMenu';
 import { NavBarMenu } from './NavBarMenu';
-import { NavBarSection } from './NavBarSection';
-import { enrichConfigItems, getActiveItem, isMatchOrChildMatch, isSearchActive, SEARCH_ITEM_ID } from './utils';
-
-const homeUrl = config.appSubUrl || '/';
+import { NavBarMenuPortalContainer } from './NavBarMenuPortalContainer';
+import { NavBarScrollContainer } from './NavBarScrollContainer';
+import { NavBarToggle } from './NavBarToggle';
+import { NavBarContext } from './context';
+import {
+  enrichConfigItems,
+  enrichWithInteractionTracking,
+  getActiveItem,
+  isMatchOrChildMatch,
+  isSearchActive,
+  SEARCH_ITEM_ID,
+} from './utils';
 
 const onOpenSearch = () => {
   locationService.partial({ search: 'open' });
 };
 
-const searchItem: NavModelItem = {
-  id: SEARCH_ITEM_ID,
-  onClick: onOpenSearch,
-  text: 'Search dashboards',
-  icon: 'search',
-};
-
-const mapStateToProps = (state: StoreState) => ({
-  navBarTree: state.navBarTree,
-});
-
-const mapDispatchToProps = {};
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
-
-export interface Props extends ConnectedProps<typeof connector> {}
-
-export const NavBarUnconnected = React.memo(({ navBarTree }: Props) => {
+export const NavBar = React.memo(() => {
+  const navBarTree = useSelector((state: StoreState) => state.navBarTree);
   const theme = useTheme2();
   const styles = getStyles(theme);
   const location = useLocation();
   const kiosk = getKioskMode();
   const [showSwitcherModal, setShowSwitcherModal] = useState(false);
+  const [menuOpen, setMenuOpen] = useState(false);
+  const [menuAnimationInProgress, setMenuAnimationInProgress] = useState(false);
+  const [menuIdOpen, setMenuIdOpen] = useState<string | undefined>(undefined);
+
   const toggleSwitcherModal = () => {
     setShowSwitcherModal(!showSwitcherModal);
   };
+
+  // Here we need to hack in a "home" and "search" NavModelItem since this is constructed in the frontend
+  const searchItem: NavModelItem = enrichWithInteractionTracking(
+    {
+      id: SEARCH_ITEM_ID,
+      onClick: onOpenSearch,
+      text: 'Search dashboards',
+      icon: 'search',
+    },
+    menuOpen
+  );
+
+  const homeItem: NavModelItem = enrichWithInteractionTracking(
+    {
+      id: 'home',
+      text: 'Home',
+      url: config.appSubUrl || '/',
+      icon: 'grafana',
+    },
+    menuOpen
+  );
+
   const navTree = cloneDeep(navBarTree);
-  const topItems = navTree.filter((item) => item.section === NavSection.Core);
-  const bottomItems = enrichConfigItems(
+
+  const coreItems = navTree
+    .filter((item) => item.section === NavSection.Core)
+    .map((item) => enrichWithInteractionTracking(item, menuOpen));
+  const pluginItems = navTree
+    .filter((item) => item.section === NavSection.Plugin)
+    .map((item) => enrichWithInteractionTracking(item, menuOpen));
+  const configItems = enrichConfigItems(
     navTree.filter((item) => item.section === NavSection.Config),
     location,
     toggleSwitcherModal
-  );
-  const activeItem = isSearchActive(location) ? searchItem : getActiveItem(navTree, location.pathname);
+  ).map((item) => enrichWithInteractionTracking(item, menuOpen));
 
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
+  const activeItem = isSearchActive(location) ? searchItem : getActiveItem(navTree, location.pathname);
 
   if (kiosk !== KioskMode.Off) {
     return null;
   }
-
   return (
-    <nav className={cx(styles.sidemenu, 'sidemenu')} data-testid="sidemenu" aria-label="Main menu">
-      <div className={styles.mobileSidemenuLogo} onClick={() => setMobileMenuOpen(!mobileMenuOpen)} key="hamburger">
-        <Icon name="bars" size="xl" />
-      </div>
+    <div className={styles.navWrapper}>
+      <nav className={cx(styles.sidemenu, 'sidemenu')} data-testid="sidemenu" aria-label="Main menu">
+        <NavBarContext.Provider
+          value={{
+            menuIdOpen: menuIdOpen,
+            setMenuIdOpen: setMenuIdOpen,
+          }}
+        >
+          <FocusScope>
+            <div className={styles.mobileSidemenuLogo} onClick={() => setMenuOpen(!menuOpen)} key="hamburger">
+              <Icon name="bars" size="xl" />
+            </div>
 
-      <NavBarSection>
-        <NavBarItemWithoutMenu label="Home" className={styles.grafanaLogo} url={homeUrl}>
-          <Branding.MenuLogo />
-        </NavBarItemWithoutMenu>
-        <NavBarItem className={styles.search} isActive={activeItem === searchItem} link={searchItem}>
-          <Icon name="search" size="xl" />
-        </NavBarItem>
-      </NavBarSection>
+            <NavBarToggle
+              className={styles.menuExpandIcon}
+              isExpanded={menuOpen}
+              onClick={() => {
+                reportInteraction('grafana_navigation_expanded');
+                setMenuOpen(true);
+              }}
+            />
 
-      <NavBarSection>
-        {topItems.map((link, index) => (
-          <NavBarItem
-            key={`${link.id}-${index}`}
-            isActive={isMatchOrChildMatch(link, activeItem)}
-            link={{ ...link, subTitle: undefined, onClick: undefined }}
-          >
-            {link.icon && <Icon name={link.icon as IconName} size="xl" />}
-            {link.img && <img src={link.img} alt={`${link.text} logo`} />}
-          </NavBarItem>
-        ))}
-      </NavBarSection>
+            <NavBarMenuPortalContainer />
 
-      <div className={styles.spacer} />
+            <NavBarItemWithoutMenu
+              elClassName={styles.grafanaLogoInner}
+              label={homeItem.text}
+              className={styles.grafanaLogo}
+              url={homeItem.url}
+              onClick={homeItem.onClick}
+            >
+              <Branding.MenuLogo />
+            </NavBarItemWithoutMenu>
 
-      <NavBarSection>
-        {bottomItems.map((link, index) => (
-          <NavBarItem
-            key={`${link.id}-${index}`}
-            isActive={isMatchOrChildMatch(link, activeItem)}
-            reverseMenuDirection
-            link={link}
-          >
-            {link.icon && <Icon name={link.icon as IconName} size="xl" />}
-            {link.img && <img src={link.img} alt={`${link.text} logo`} />}
-          </NavBarItem>
-        ))}
-      </NavBarSection>
+            <NavBarScrollContainer>
+              <ul className={styles.itemList}>
+                <NavBarItem className={styles.search} isActive={activeItem === searchItem} link={searchItem} />
 
+                {coreItems.map((link, index) => (
+                  <NavBarItem
+                    key={`${link.id}-${index}`}
+                    isActive={isMatchOrChildMatch(link, activeItem)}
+                    link={{ ...link, subTitle: undefined }}
+                  />
+                ))}
+
+                {pluginItems.length > 0 &&
+                  pluginItems.map((link, index) => (
+                    <NavBarItem
+                      key={`${link.id}-${index}`}
+                      isActive={isMatchOrChildMatch(link, activeItem)}
+                      link={link}
+                    />
+                  ))}
+
+                {configItems.map((link, index) => (
+                  <NavBarItem
+                    key={`${link.id}-${index}`}
+                    isActive={isMatchOrChildMatch(link, activeItem)}
+                    reverseMenuDirection
+                    link={link}
+                    className={cx({ [styles.verticalSpacer]: index === 0 })}
+                  />
+                ))}
+              </ul>
+            </NavBarScrollContainer>
+          </FocusScope>
+        </NavBarContext.Provider>
+      </nav>
       {showSwitcherModal && <OrgSwitcher onDismiss={toggleSwitcherModal} />}
-      {mobileMenuOpen && (
-        <NavBarMenu
-          activeItem={activeItem}
-          navItems={[searchItem, ...topItems, ...bottomItems]}
-          onClose={() => setMobileMenuOpen(false)}
-        />
+      {(menuOpen || menuAnimationInProgress) && (
+        <div className={styles.menuWrapper}>
+          <NavBarMenu
+            activeItem={activeItem}
+            isOpen={menuOpen}
+            setMenuAnimationInProgress={setMenuAnimationInProgress}
+            navItems={[homeItem, searchItem, ...coreItems, ...pluginItems, ...configItems]}
+            onClose={() => setMenuOpen(false)}
+          />
+        </div>
       )}
-    </nav>
+    </div>
   );
 });
 
-NavBarUnconnected.displayName = 'NavBar';
-
-export const NavBar = connector(NavBarUnconnected);
+NavBar.displayName = 'NavBar';
 
 const getStyles = (theme: GrafanaTheme2) => ({
-  search: css`
-    display: none;
-    margin-top: ${theme.spacing(5)};
+  navWrapper: css({
+    position: 'relative',
+    display: 'flex',
 
-    ${theme.breakpoints.up('md')} {
-      display: block;
-    }
-  `,
-  sidemenu: css`
-    display: flex;
-    flex-direction: column;
-    position: fixed;
-    z-index: ${theme.zIndex.sidemenu};
+    '.sidemenu-hidden &': {
+      display: 'none',
+    },
+  }),
+  sidemenu: css({
+    label: 'sidemenu',
+    display: 'flex',
+    flexDirection: 'column',
+    backgroundColor: theme.colors.background.primary,
+    zIndex: theme.zIndex.sidemenu,
+    padding: `${theme.spacing(1)} 0`,
+    position: 'relative',
+    width: theme.components.sidemenu.width,
+    borderRight: `1px solid ${theme.colors.border.weak}`,
 
-    ${theme.breakpoints.up('md')} {
-      background: ${theme.colors.background.primary};
-      border-right: 1px solid ${theme.components.panel.borderColor};
-      padding: 0 0 ${theme.spacing(1)} 0;
-      position: relative;
-      width: ${theme.components.sidemenu.width}px;
-    }
+    [theme.breakpoints.down('md')]: {
+      height: theme.spacing(7),
+      position: 'fixed',
+      paddingTop: '0px',
+      backgroundColor: 'inherit',
+      borderRight: 0,
+    },
+  }),
+  mobileSidemenuLogo: css({
+    alignItems: 'center',
+    cursor: 'pointer',
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    padding: theme.spacing(2),
 
-    .sidemenu-hidden & {
-      display: none;
-    }
-  `,
-  grafanaLogo: css`
-    display: none;
-    img {
-      height: ${theme.spacing(3.5)};
-      width: ${theme.spacing(3.5)};
-    }
+    [theme.breakpoints.up('md')]: {
+      display: 'none',
+    },
+  }),
+  itemList: css({
+    backgroundColor: 'inherit',
+    display: 'flex',
+    flexDirection: 'column',
+    height: '100%',
 
-    ${theme.breakpoints.up('md')} {
-      align-items: center;
-      display: flex;
-      justify-content: center;
-    }
-  `,
-  mobileSidemenuLogo: css`
-    align-items: center;
-    cursor: pointer;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    padding: ${theme.spacing(2)};
+    [theme.breakpoints.down('md')]: {
+      visibility: 'hidden',
+    },
+  }),
+  grafanaLogo: css({
+    alignItems: 'stretch',
+    display: 'flex',
+    flexShrink: 0,
+    height: theme.spacing(6),
+    justifyContent: 'stretch',
 
-    ${theme.breakpoints.up('md')} {
-      display: none;
-    }
-  `,
-  spacer: css`
-    flex: 1;
-  `,
+    [theme.breakpoints.down('md')]: {
+      visibility: 'hidden',
+    },
+  }),
+  grafanaLogoInner: css({
+    alignItems: 'center',
+    display: 'flex',
+    height: '100%',
+    justifyContent: 'center',
+    width: '100%',
+
+    '> div': {
+      height: 'auto',
+      width: 'auto',
+    },
+  }),
+  search: css({
+    display: 'none',
+    marginTop: 0,
+
+    [theme.breakpoints.up('md')]: {
+      display: 'grid',
+    },
+  }),
+  verticalSpacer: css({
+    marginTop: 'auto',
+  }),
+  hideFromMobile: css({
+    [theme.breakpoints.down('md')]: {
+      display: 'none',
+    },
+  }),
+  menuWrapper: css({
+    position: 'fixed',
+    display: 'grid',
+    gridAutoFlow: 'column',
+    height: '100%',
+    zIndex: theme.zIndex.sidemenu,
+  }),
+  menuExpandIcon: css({
+    position: 'absolute',
+    top: '43px',
+    right: '0px',
+    transform: `translateX(50%)`,
+  }),
+  menuPortalContainer: css({
+    zIndex: theme.zIndex.sidemenu,
+  }),
 });
