+++ public/app/core/components/NavBar/NavBarItem.tsx
@@ -1,7 +1,7 @@
 import { css, cx } from '@emotion/css';
 import { useLingui } from '@lingui/react';
 import { Item } from '@react-stately/collections';
-import React, { ReactNode } from 'react';
+import React from 'react';
 
 import { GrafanaTheme2, locationUtil, NavMenuItemType, NavModelItem } from '@grafana/data';
 import { locationService } from '@grafana/runtime';
@@ -9,31 +9,24 @@ import { IconName, useTheme2 } from '@grafana/ui';
 
 import { NavBarItemMenu } from './NavBarItemMenu';
 import { NavBarItemMenuTrigger } from './NavBarItemMenuTrigger';
-import { getNavBarItemWithoutMenuStyles, NavBarItemWithoutMenu } from './NavBarItemWithoutMenu';
+import { getNavBarItemWithoutMenuStyles } from './NavBarItemWithoutMenu';
 import { NavBarMenuItem } from './NavBarMenuItem';
+import { useNavBarContext } from './context';
 import menuItemTranslations from './navBarItem-translations';
 import { getNavModelItemKey } from './utils';
 
 export interface Props {
   isActive?: boolean;
-  children: ReactNode;
   className?: string;
   reverseMenuDirection?: boolean;
-  showMenu?: boolean;
   link: NavModelItem;
 }
 
-const NavBarItem = ({
-  isActive = false,
-  children,
-  className,
-  reverseMenuDirection = false,
-  showMenu = true,
-  link,
-}: Props) => {
+const NavBarItem = ({ isActive = false, className, reverseMenuDirection = false, link }: Props) => {
   const { i18n } = useLingui();
   const theme = useTheme2();
   const menuItems = link.children ?? [];
+  const { menuIdOpen } = useNavBarContext();
 
   // Spreading `menuItems` here as otherwise we'd be mutating props
   const menuItemsSorted = reverseMenuDirection ? [...menuItems].reverse() : menuItems;
@@ -51,24 +44,28 @@ const NavBarItem = ({
 
   const onNavigate = (item: NavModelItem) => {
     const { url, target, onClick } = item;
-    if (!url) {
-      onClick?.();
-      return;
-    }
+    onClick?.();
 
-    if (!target && url.startsWith('/')) {
-      locationService.push(locationUtil.stripBaseFromUrl(url));
-    } else {
-      window.open(url, target);
+    if (url) {
+      if (!target && url.startsWith('/')) {
+        locationService.push(locationUtil.stripBaseFromUrl(url));
+      } else {
+        window.open(url, target);
+      }
     }
   };
 
   const translationKey = link.id && menuItemTranslations[link.id];
   const linkText = translationKey ? i18n._(translationKey) : link.text;
 
-  return showMenu ? (
-    <li className={cx(styles.container, className)}>
-      <NavBarItemMenuTrigger item={section} isActive={isActive} label={linkText}>
+  return (
+    <li className={cx(styles.container, { [styles.containerHover]: section.id === menuIdOpen }, className)}>
+      <NavBarItemMenuTrigger
+        item={section}
+        isActive={isActive}
+        label={linkText}
+        reverseMenuDirection={reverseMenuDirection}
+      >
         <NavBarItemMenu
           items={items}
           reverseMenuDirection={reverseMenuDirection}
@@ -80,31 +77,19 @@ const NavBarItem = ({
           {(item: NavModelItem) => {
             const translationKey = item.id && menuItemTranslations[item.id];
             const itemText = translationKey ? i18n._(translationKey) : item.text;
-
-            if (item.menuItemType === NavMenuItemType.Section) {
-              return (
-                <Item key={getNavModelItemKey(item)} textValue={item.text}>
-                  <NavBarMenuItem
-                    target={item.target}
-                    text={itemText}
-                    url={item.url}
-                    onClick={item.onClick}
-                    styleOverrides={styles.header}
-                  />
-                </Item>
-              );
-            }
+            const isSection = item.menuItemType === NavMenuItemType.Section;
+            const icon = item.showIconInNavbar && !isSection ? (item.icon as IconName) : undefined;
 
             return (
               <Item key={getNavModelItemKey(item)} textValue={item.text}>
                 <NavBarMenuItem
-                  isDivider={item.divider}
-                  icon={item.icon as IconName}
-                  onClick={item.onClick}
+                  isDivider={!isSection && item.divider}
+                  icon={icon}
                   target={item.target}
                   text={itemText}
                   url={item.url}
-                  styleOverrides={styles.item}
+                  onClick={item.onClick}
+                  styleOverrides={cx(styles.primaryText, { [styles.header]: isSection })}
                 />
               </Item>
             );
@@ -112,18 +97,6 @@ const NavBarItem = ({
         </NavBarItemMenu>
       </NavBarItemMenuTrigger>
     </li>
-  ) : (
-    <NavBarItemWithoutMenu
-      label={link.text}
-      className={className}
-      isActive={isActive}
-      url={link.url}
-      onClick={link.onClick}
-      target={link.target}
-      highlightText={link.highlightText}
-    >
-      {children}
-    </NavBarItemWithoutMenu>
   );
 };
 
@@ -131,16 +104,19 @@ export default NavBarItem;
 
 const getStyles = (theme: GrafanaTheme2, adjustHeightForBorder: boolean, isActive?: boolean) => ({
   ...getNavBarItemWithoutMenuStyles(theme, isActive),
-  header: css`
-    color: ${theme.colors.text.primary};
-    height: ${theme.components.sidemenu.width - (adjustHeightForBorder ? 2 : 1)}px;
-    font-size: ${theme.typography.h4.fontSize};
-    font-weight: ${theme.typography.h4.fontWeight};
-    padding: ${theme.spacing(1)} ${theme.spacing(2)};
-    white-space: nowrap;
-    width: 100%;
-  `,
-  item: css`
-    color: ${theme.colors.text.primary};
-  `,
+  containerHover: css({
+    backgroundColor: theme.colors.action.hover,
+    color: theme.colors.text.primary,
+  }),
+  primaryText: css({
+    color: theme.colors.text.primary,
+  }),
+  header: css({
+    height: `calc(${theme.spacing(6)} - ${adjustHeightForBorder ? 2 : 1}px)`,
+    fontSize: theme.typography.h4.fontSize,
+    fontWeight: theme.typography.h4.fontWeight,
+    padding: `${theme.spacing(1)} ${theme.spacing(2)}`,
+    whiteSpace: 'nowrap',
+    width: '100%',
+  }),
 });
